Implement `PackageJson` in dedicated file
